requirements.txt
discord.py
random
python-dotenv